+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Reorganized configuration entries for more clarity
 * __Please revise your configuration once after installing this version__
 * Added saving of local scaling into blueprints. Note that the scale is not saved/synced by the vanilla game for most of the prefabs, so an additional mod for piece scaling should be installed.
+* Added saving of item stand and armor stand items. Loadouts only get placed when using "Direct Build".
 * Added the ability to change blueprint categories via the Marketplace GUI
 * Removed the blueprint "Quick Select" tool, use the add/remove/edit selection tools instead
 * Added temporary capturing of blueprints from the "Edit Selection" tool at runtime without saving a file for it (copy+paste)